UI Finalzing

+++ Game/vs2012 base project/msvc/JewelryItem.h
@@ -3,7 +3,7 @@
 class JewelryItem :public Item
 {
 public:
-	JewelryItem(int itemId,int x, int y) :Item(itemId,x,y){}
+	JewelryItem(int itemId, float x, float y) :Item(itemId,x,y){}
 	void update();
 	void draw(King::Engine& engine);
 	void move(float x , float y);
